+++ test/functional/android/android/device_test.rb
@@ -8,6 +8,9 @@ class AppiumLibCoreTest
       def setup
         @@core ||= ::Appium::Core.for(Caps.android)
         @driver = @@core.start_driver
+
+        require 'pry'
+        binding.pry
       end
 
       def teardown